change CreateArtifactoryDetailsByFlags func for rt and ds
